server/requirements.txt
fastapi
uvicorn
langchain-community
langchain-groq
langchain-google-genai
chromadb
sentence-transformers
pypdf
python-dotenv
python-multipart
aiofiles
tiktoken
